+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Keep it human-readable, your future self will thank you!
 
 ### Added
 
+- fix: bug in color when plotting isolated nodes
+
+### Added
+
 - Add anemoi-transform link to documentation
 
 ### Added
@@ -20,6 +24,8 @@ Keep it human-readable, your future self will thank you!
 - feat: New class, KNNAreaMaskBuilder, to specify Area of Interest (AOI) based on a set of nodes. (#30)
 - feat: New node builder classes, LimitedAreaXXXXXNodes, to create nodes within an Area of Interest (AOI). (#30)
 - feat: Expanded MultiScaleEdges to support multi-scale connections in limited area graphs. (#30)
+- feat: New class StretchedTriNodes to create a stretched mesh. (#51)
+- feat: Expanded MultiScaleEdges to support multi-scale connections in stretched graphs. (#51)
 
 ### Changed
 - ci: small fixes and updates pre-commit, downsteam-ci (#49)